# Conflicts:
#	modules/skybrowser/skybrowsermodule_lua.inl

+++ modules/globebrowsing/src/layermanager.cpp
@@ -2,7 +2,7 @@
  *                                                                                       *
  * OpenSpace                                                                             *
  *                                                                                       *
- * Copyright (c) 2014-2022                                                               *
+ * Copyright (c) 2014-2023                                                               *
  *                                                                                       *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
  * software and associated documentation files (the "Software"), to deal in the Software *
@@ -145,7 +145,7 @@ std::array<LayerGroup*, LayerManager::NumLayerGroups> LayerManager::layerGroups(
     ZoneScoped
 
     std::array<LayerGroup*, NumLayerGroups> res = {};
-    for (int i = 0; i < NumLayerGroups; ++i) {
+    for (size_t i = 0; i < NumLayerGroups; ++i) {
         res[i] = _layerGroups[i].get();
     }
     return res;
